Interfaces into Model
