+++ projects/swimlane/ngx-charts/src/lib/pie-chart/pie-grid.component.ts
@@ -17,12 +17,11 @@ import { BaseChartComponent } from '../common/base-chart.component';
 import { trimLabel } from '../common/trim-label.helper';
 import { gridLayout } from '../common/grid-layout.helper';
 import { formatLabel } from '../common/label.helper';
-import { DataItem } from '../models/chart-data.model';
+import { DataItem, PieGridData } from '../models/chart-data.model';
 import { PlacementTypes } from '../common/tooltip/position';
 import { StyleTypes } from '../common/tooltip/style.type';
 import { ViewDimensions } from '../common/types/view-dimension.interface';
 import { ScaleType } from '../common/types/scale-type.enum';
-import { PieGridData } from './pie-grid-series.component';
 
 @Component({
   selector: 'ngx-charts-pie-grid',
